Michael Schwingen <rincewind@discworld.dascon.de>  enhanced non-CFI flash support
git-svn-id: svn://svn.berlios.de/openocd/trunk@1469 b42882b7-edfa-0310-969c-e2dbd0fdcd60

+++ src/flash/cfi.h
@@ -26,7 +26,7 @@
 typedef struct cfi_flash_bank_s
 {
 	working_area_t *write_algorithm;
-	
+
 
 	int x16_as_x8;
 	int jedec_probe;
@@ -59,7 +59,7 @@ typedef struct cfi_flash_bank_s
 	u8 chip_erase_timeout_max;
 
 	/* flash geometry */
-	u8 dev_size;
+	u32 dev_size;
 	u16 interface_desc;
 	u16 max_buf_write_size;
 	u8 num_erase_regions;
